Move BDD core and BDDIpSpaceSpecializer to common
