+++ projects/batfish-common-protocol/pom.xml
@@ -207,6 +207,11 @@
     </dependency>
 
     <dependency>
+      <groupId>net.sf.javabdd</groupId>
+      <artifactId>javabdd</artifactId>
+    </dependency>
+
+    <dependency>
       <groupId>org.apache.commons</groupId>
       <artifactId>commons-configuration2</artifactId>
     </dependency>
