+++ projects/batfish-common-protocol/src/main/java/org/batfish/common/bdd/BDDUtils.java
@@ -1,4 +1,4 @@
-package org.batfish.symbolic.bdd;
+package org.batfish.common.bdd;
 
 import net.sf.javabdd.BDD;
 import net.sf.javabdd.BDDFactory;
