+++ projects/batfish-common-protocol/src/main/java/org/batfish/common/bdd/MemoizedIpSpaceToBDD.java
@@ -1,4 +1,4 @@
-package org.batfish.symbolic.bdd;
+package org.batfish.common.bdd;
 
 import com.google.common.annotations.VisibleForTesting;
 import java.util.IdentityHashMap;
