+++ projects/batfish-common-protocol/src/test/java/org/batfish/common/bdd/BDDIpSpaceSpecializerTest.java
@@ -1,4 +1,4 @@
-package org.batfish.z3;
+package org.batfish.common.bdd;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
@@ -14,9 +14,6 @@ import org.batfish.datamodel.IpWildcard;
 import org.batfish.datamodel.IpWildcardSetIpSpace;
 import org.batfish.datamodel.Prefix;
 import org.batfish.datamodel.UniverseIpSpace;
-import org.batfish.symbolic.bdd.BDDInteger;
-import org.batfish.symbolic.bdd.BDDUtils;
-import org.batfish.symbolic.bdd.IpSpaceToBDD;
 import org.junit.Before;
 import org.junit.Test;
