+++ projects/batfish/src/main/java/org/batfish/symbolic/abstraction/BatfishCompressor.java
@@ -18,6 +18,7 @@ import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import org.batfish.common.bdd.BDDPacket;
 import org.batfish.common.plugin.IBatfish;
 import org.batfish.datamodel.BgpActivePeerConfig;
 import org.batfish.datamodel.Configuration;
@@ -46,7 +47,6 @@ import org.batfish.datamodel.routing_policy.statement.Statements;
 import org.batfish.symbolic.Graph;
 import org.batfish.symbolic.GraphEdge;
 import org.batfish.symbolic.Protocol;
-import org.batfish.symbolic.bdd.BDDPacket;
 import org.batfish.symbolic.collections.Table2;
 
 /*
